% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianETAS-package.R
\docType{package}
\name{bayesianETAS}
\alias{bayesianETAS}
\alias{bayesianETAS-package}
\title{Bayesian estimation of the ETAS model for earthquake occurrences}
\description{
Bayesian estimation of the ETAS model for earthquake occurrences
}
\author{
Gordon J Ross \email{gordon@gordonjross.co.uk}
}
\references{
Gordon J. Ross - Bayesian Estimation of the ETAS Model for Earthquake Occurrences (2016), available from http://www.gordonjross.co.uk/bayesianetas.pdf
}

